+++ milestone05/src/arch/x86_64/ps2.c
@@ -30,14 +30,13 @@ uint8_t ps2_poll_read(uint16_t port){
    return inb(port);
 }
 
-int8_t ps2_read(uint16_t port){
-   uint8_t status = inb(PS2_STATUS);
+int16_t ps2_read(uint16_t port){
+   uint16_t status = inb(PS2_STATUS);
 
    if (status & PS2_STATUS_OUTPUT)
       return inb(port);
    else
       return -1;
-
 }
