+++ milestone05/src/arch/x86_64/ps2.h
@@ -5,7 +5,7 @@
 
 void initialize_PS2_controller(void);
 uint8_t ps2_poll_read(uint16_t port);
-int8_t ps2_read(uint16_t port);
+int16_t ps2_read(uint16_t port);
 void ps2_poll_write(uint16_t port, uint8_t val);
 
 #endif
